new registration component and simplified login
