+++ src/Components/Register/RegistrationForm.css
@@ -1,0 +1,30 @@
+@import url("https://fonts.googleapis.com/css2?family=Fira+Sans:wght@100;300;400;500;600;700&display=swap");
+
+.form-container {
+    padding-bottom: 100px;
+    margin: 0px auto;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.form-container a,
+.form-container a:visited {
+  color: white;
+}
+
+.log-form {
+    /* border: 1px solid red; */
+    width: 95%;
+    max-width: 550px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 30px 0;
+    border-radius: 5px;
+    margin-top: 125px;
+    box-shadow: var(--box-shadow);
+    background-color: var(--color-primary);
+    background-color: #302c2c;
+    color: white;
+  }
